YTDB-510: Fix code

+++ core/src/main/java/com/jetbrains/youtrackdb/internal/core/storage/collection/v2/PaginatedCollectionV2.java
@@ -350,7 +350,7 @@ public final class PaginatedCollectionV2 extends PaginatedCollection {
   @Override
   public PhysicalPosition createRecord(
       final byte[] content,
-      final int recordVersion,
+      final long recordVersion,
       final byte recordType,
       final PhysicalPosition allocatedPosition,
       final AtomicOperation atomicOperation) {
@@ -466,7 +466,7 @@ public final class PaginatedCollectionV2 extends PaginatedCollection {
       final byte[] content,
       final int len,
       final byte recordType,
-      final int recordVersion,
+      final long recordVersion,
       final long collectionPosition,
       final long nextRecordPointer,
       final AtomicOperation atomicOperation,
@@ -908,7 +908,7 @@ public final class PaginatedCollectionV2 extends PaginatedCollection {
   public void updateRecord(
       final long collectionPosition,
       final byte[] content,
-      final int recordVersion,
+      final long recordVersion,
       final byte recordType,
       final AtomicOperation atomicOperation) {
     executeInsideComponentOperation(
@@ -1046,7 +1046,7 @@ public final class PaginatedCollectionV2 extends PaginatedCollection {
   }
 
   @Override
-  public void updateRecordVersion(long collectionPosition, int recordVersion,
+  public void updateRecordVersion(long collectionPosition, long recordVersion,
       AtomicOperation atomicOperation) {
     executeInsideComponentOperation(
         atomicOperation,
@@ -1464,7 +1464,7 @@ public final class PaginatedCollectionV2 extends PaginatedCollection {
   }
 
   private static PhysicalPosition createPhysicalPosition(
-      final byte recordType, final long collectionPosition, final int version) {
+      final byte recordType, final long collectionPosition, final long version) {
     final var physicalPosition = new PhysicalPosition();
     physicalPosition.recordType = recordType;
     physicalPosition.recordSize = -1;
